#3 fix: redis 관련 메서드는 RedisService에서 구현하도록 수정
